Build Agda and Idris in the same Cabal file
Makes it possible to build both Agda and Idris, a full Idris build,
using the same project and dependencies. Idris uses a custom build which
had to be adapted to a new folder structure.

This makes it possible to do a full Idris build and reuse both the
parser, typechecker and elaborator in my project.